re-factored dropdown parent to child component
